# internal/apiserver/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.CreateRoleRequest:
    properties:
      description:
        type: string
      display_name:
        type: string
      name:
        type: string
    required:
    - display_name
    - name
    type: object
...
  github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.ListResponse:
    properties:
      code:
        type: integer
      data: {}
      limit:
        type: integer
      message:
        type: string
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.RoleResponse:
    properties:
      description:
        type: string
      display_name:
        type: string
      id:
        type: string
      name:
        type: string
      tenant_id:
        type: string
    type: object
...
  github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.UpdateRoleRequest:
    properties:
      description:
        type: string
      display_name:
        type: string
    type: object
...
  github_com_FangcunMount_iam-contracts_internal_apiserver_interface_uc_restful_request.GuardianGrantRequest:
    properties:
      childId:
        type: string
      relation:
        enum:
        - self
        - parent
        - grandparent
        - other
        type: string
      userId:
        type: string
    required:
    - childId
    - relation
    type: object
...
  github_com_FangcunMount_iam-contracts_internal_apiserver_interface_uc_restful_response.GuardianshipPageResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_FangcunMount_iam-contracts_internal_apiserver_interface_uc_restful_response.GuardianshipResponse'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  github_com_FangcunMount_iam-contracts_internal_apiserver_interface_uc_restful_response.GuardianshipResponse:
    properties:
      childId:
        type: string
      id:
        type: integer
      relation:
        type: string
      revokedAt:
        type: string
      since:
        type: string
      userId:
        type: string
    type: object
...
  github_com_FangcunMount_iam-contracts_pkg_core.ErrResponse:
    properties:
      code:
        description: Code 定义了业务错误代码
        type: integer
      message:
        description: |-
          Message 包含此消息的详细信息
          此消息适合暴露给外部
        type: string
      reference:
        description: Reference 返回参考文档，可能有助于解决此错误
        type: string
    type: object
host: iam.fangcunmount.cn
info:
  contact:
    email: support@fangcunmount.cn
    name: API Support
    url: https://github.com/FangcunMount/iam-contracts
  description: IAM 系统 REST API 文档，包含认证(Authentication)、授权(Authorization)、身份管理(Identity)和身份提供商(IDP)模块
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://iam.fangcunmount.cn/terms
  title: IAM API Documentation
  version: 1.0.0
paths:
...
  /authz/roles:
    get:
      parameters:
      - default: 0
        description: 偏移量
        in: query
        name: offset
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.ListResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.RoleResponse'
                  type: array
              type: object
      summary: 列出角色
      tags:
      - Authorization-Roles
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建角色请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.RoleResponse'
              type: object
      summary: 创建角色
      tags:
      - Authorization-Roles
  /authz/roles/{id}:
    delete:
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.Response'
      summary: 删除角色
      tags:
      - Authorization-Roles
    get:
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.RoleResponse'
              type: object
      summary: 获取角色详情
      tags:
      - Authorization-Roles
    put:
      consumes:
      - application/json
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新角色请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.RoleResponse'
              type: object
      summary: 更新角色
      tags:
      - Authorization-Roles
...
  /identity/guardians:
    get:
      consumes:
      - application/json
      description: 查询当前登录用户视角下的监护关系；省略 user_id 时默认当前用户，按 child_id 查询前要求当前用户对该儿童存在活跃监护关系。
      parameters:
      - description: 当前用户 ID；省略时默认当前登录用户，不支持跨用户查询
        in: query
        name: user_id
        type: string
      - description: 儿童 ID；仅当当前用户对该儿童存在活跃监护关系时可查询
        in: query
        name: child_id
        type: string
      - description: 是否仅查询活跃的监护关系
        in: query
        name: active
        type: boolean
      - default: 0
        description: 偏移量
        in: query
        name: offset
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            $ref: '#/definitions/github_com_FangcunMount_iam-contracts_internal_apiserver_interface_uc_restful_response.GuardianshipPageResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/github_com_FangcunMount_iam-contracts_pkg_core.ErrResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/github_com_FangcunMount_iam-contracts_pkg_core.ErrResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_FangcunMount_iam-contracts_pkg_core.ErrResponse'
      security:
      - BearerAuth: []
      summary: 查询监护关系
      tags:
      - Identity-Guardianship
  /identity/guardians/grant:
    post:
      consumes:
      - application/json
      description: 将当前登录用户设置为儿童的监护人；如传 userId，必须与当前登录用户一致
      parameters:
      - description: 授予监护请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_FangcunMount_iam-contracts_internal_apiserver_interface_uc_restful_request.GuardianGrantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 授予成功
          schema:
            $ref: '#/definitions/github_com_FangcunMount_iam-contracts_internal_apiserver_interface_uc_restful_response.GuardianshipResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/github_com_FangcunMount_iam-contracts_pkg_core.ErrResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/github_com_FangcunMount_iam-contracts_pkg_core.ErrResponse'
        "409":
          description: 监护关系已存在
          schema:
            $ref: '#/definitions/github_com_FangcunMount_iam-contracts_pkg_core.ErrResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_FangcunMount_iam-contracts_pkg_core.ErrResponse'
      security:
      - BearerAuth: []
      summary: 授予监护关系
      tags:
      - Identity-Guardianship
...
securityDefinitions:
  BearerAuth:
    description: 'JWT 认证令牌，格式: Bearer {access_token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...